Refactor --nodename to be --hostname, and add-in a user-defined node name.
This refactoring allows pg_auto_failover to manage a user-defined public
node name for the nodes separate from the hostname used to connect to the
Postgres service on nodes. It appears that people like naming things.

Next-up in this PR: allowing the edit the metadata of a node at any point,
in particular change the name or hostname or port at pg_autoctl run, for
easy k8s integration (--hostname may change at each pod reboot).