Implement pg_autoctl set node metadata.
This allows changing the registration of a node to a new name, hostname, and
port if needed. That can be done at runtime, except for the Postgres port
that we can't change without restarting Postgres anyway.

At the moment the new command is hidden in the PG_AUTOCTL_DEBUG facility,
this should change with the exposure of the other get/set commands, or could
be done from within this PR.

Should fix #213, explicit QA still needed at this point.